split layernorm factory file into 2
