+++ ttnn/cpp/ttnn/operations/normalization/layernorm/device/layernorm_common.hpp
@@ -1,0 +1,16 @@
+// SPDX-FileCopyrightText: © 2023 Tenstorrent Inc.
+//
+// SPDX-License-Identifier: Apache-2.0
+
+#pragma once
+
+#include <optional>
+#include <tt-metalium/host_api.hpp>
+#include "ttnn/tensor/tensor.hpp"
+
+namespace ttnn::operations::normalization {
+
+std::pair<std::optional<tt::tt_metal::Tensor>, uint32_t> create_reciprocal_tensor_if_needed(
+    tt::tt_metal::IDevice* device, uint32_t W, const tt::tt_metal::CoreRangeSet& cores, const bool use_welford);
+
+}  // namespace ttnn::operations::normalization
